Update qa-system.py
The main updates I made where adding code so that where are when questions with common formats are answered.  I did this by adding more regex to the code.  I think that before sunday we just need to focus on adding more regex's to the code so that our system can flag answers for different wikipedia formats.  I left the log in my py file so that you all can see the types of questions it can currently answer.